Restrict with_exprt to exactly three operands
We cannot build a C++ API for the previous >= 3, odd-number operand
variant. Alas, we ended up with various places in the code base silently
ignoring additional operands, which led to wrong verification results in
Kani as recent changes in CBMC made increasing use of the value set
(which is one of the places that silently ignored additional operands).

Resolves: #7272