MacOS CI job: do not re-install CMake
This causes `brew` to complain about conflicting taps due to pinning
that was done (by GitHub) to alleviate a possible build failure by CMake
4.x. See https://github.com/actions/runner-images/issues/12912, and
should be fixed (by GitHub) already, but apparently that's not uniformly
the case. The rollout of CMake 4.x is also blocked by those issues (cf.
https://github.com/actions/runner-images/issues/12934).

See
https://github.com/diffblue/cbmc/actions/runs/17673397336/job/50229845951?pr=8646
for an example of a failing run on our end.