Makefile builds: do not inherit SRC from environment in big-int
In just `src/big-int` we were including `common` before defining `SRC`,
which meant a possible environment-defined value of `SRC` was being used
to construct some rules. The possible reason for having the `include`
earlier in just that file was that we were lacking a rule for `.cc`
files, the absence of which caused `make` to fail when sourcing `common`
after setting `SRC`.

Both fixes together now ensure that even an environment with `SRC` set
won't cause builds to fail in strange ways.

Fixes: #4834