CONTRACTS: bugfix: make ptr_pred_ctx_reset_call call reset instead of init function
Fixes a copy and paste bug that results in a soundness bug.

The method ptr_pred_ctx_reset_call generated a call to the INIT
function instead of the intended RESET function, which resulted
in separation checks for the __CPROVER_is fresh memory
predicate failing to work  across function calls between requires
and ensures clauses.